Feat: moved newContacts to contacts folder
